cmd/zelda: redirect PE imports to ELF dynamic symbols
This commit adds preliminary support for translating
PE import entries to ELF dynamic symbols.

More specifically, for each imported function the
address of the functions .plt entry is added to
the corresponding PE import entry; that is to the
Import Address Table (IAT) entry of that import.

@7i with this commit, we are ready to start implementing
functions in hallondricka.